Tons 'o DB refactoring
