+++ modules/database.js
@@ -3,7 +3,6 @@
  * each individual database module.
  */
 
-require("./database/dbConnSetup");
 
 exports.addQuestions = require("./database/addQuestions").addQuestions;
 exports.createSurvey = require("./database/createSurvey").createSurvey;
@@ -15,7 +14,6 @@ exports.removeQuestion = require("./database/removeQuestion").removeQuestion;
 exports.addQuestions = require("./database/addQuestions").addQuestions;
 exports.addAnswer = require("./database/addAnswer").addAnswer;
 exports.removeAnswer = require("./database/removeAnswer").removeAnswer;
-exports.pgConnect = require("./database/pgConnect").pgConnect;
